move constants into static class EventReporterConstants
