+++ azkaban-exec-server/src/main/java/azkaban/execapp/FlowRunner.java
@@ -15,41 +15,41 @@
  */
 package azkaban.execapp;
 
-import static azkaban.Constants.ATTEMPT_ID;
-import static azkaban.Constants.AZ_HOST;
-import static azkaban.Constants.AZ_WEBSERVER;
+import static azkaban.Constants.EventReporterConstants.ATTEMPT_ID;
+import static azkaban.Constants.EventReporterConstants.AZ_HOST;
+import static azkaban.Constants.EventReporterConstants.AZ_WEBSERVER;
 import static azkaban.Constants.ConfigurationKeys.AZKABAN_EVENT_REPORTING_PROPERTIES_TO_PROPAGATE;
 import static azkaban.Constants.ConfigurationKeys.AZKABAN_SERVER_HOST_NAME;
 import static azkaban.Constants.ConfigurationKeys.AZKABAN_WEBSERVER_EXTERNAL_HOSTNAME;
-import static azkaban.Constants.END_TIME;
-import static azkaban.Constants.EXECUTION_ID;
-import static azkaban.Constants.EXECUTOR_TYPE;
-import static azkaban.Constants.FAILED_JOB_ID;
-import static azkaban.Constants.FAILURE_MESSAGE;
-import static azkaban.Constants.FLOW_KILL_DURATION;
-import static azkaban.Constants.FLOW_NAME;
-import static azkaban.Constants.FLOW_PAUSE_DURATION;
-import static azkaban.Constants.FLOW_PREPARATION_DURATION;
-import static azkaban.Constants.FLOW_STATUS;
-import static azkaban.Constants.FLOW_VERSION;
-import static azkaban.Constants.JOB_ID;
-import static azkaban.Constants.JOB_KILL_DURATION;
-import static azkaban.Constants.JOB_PROXY_USER;
-import static azkaban.Constants.JOB_STATUS;
-import static azkaban.Constants.JOB_TYPE;
-import static azkaban.Constants.MODIFIED_BY;
-import static azkaban.Constants.PROJECT_FILE_NAME;
-import static azkaban.Constants.PROJECT_FILE_UPLOADER_IP_ADDR;
-import static azkaban.Constants.PROJECT_FILE_UPLOAD_TIME;
-import static azkaban.Constants.PROJECT_FILE_UPLOAD_USER;
-import static azkaban.Constants.PROJECT_NAME;
-import static azkaban.Constants.QUEUE_DURATION;
-import static azkaban.Constants.SLA_OPTIONS;
-import static azkaban.Constants.START_TIME;
-import static azkaban.Constants.SUBMIT_TIME;
-import static azkaban.Constants.SUBMIT_USER;
-import static azkaban.Constants.VERSION;
-import static azkaban.Constants.VERSION_SET;
+import static azkaban.Constants.EventReporterConstants.END_TIME;
+import static azkaban.Constants.EventReporterConstants.EXECUTION_ID;
+import static azkaban.Constants.EventReporterConstants.EXECUTOR_TYPE;
+import static azkaban.Constants.EventReporterConstants.FAILED_JOB_ID;
+import static azkaban.Constants.EventReporterConstants.FAILURE_MESSAGE;
+import static azkaban.Constants.EventReporterConstants.FLOW_KILL_DURATION;
+import static azkaban.Constants.EventReporterConstants.FLOW_NAME;
+import static azkaban.Constants.EventReporterConstants.FLOW_PAUSE_DURATION;
+import static azkaban.Constants.EventReporterConstants.FLOW_PREPARATION_DURATION;
+import static azkaban.Constants.EventReporterConstants.FLOW_STATUS;
+import static azkaban.Constants.EventReporterConstants.FLOW_VERSION;
+import static azkaban.Constants.EventReporterConstants.JOB_ID;
+import static azkaban.Constants.EventReporterConstants.JOB_KILL_DURATION;
+import static azkaban.Constants.EventReporterConstants.JOB_PROXY_USER;
+import static azkaban.Constants.EventReporterConstants.JOB_STATUS;
+import static azkaban.Constants.EventReporterConstants.JOB_TYPE;
+import static azkaban.Constants.EventReporterConstants.MODIFIED_BY;
+import static azkaban.Constants.EventReporterConstants.PROJECT_FILE_NAME;
+import static azkaban.Constants.EventReporterConstants.PROJECT_FILE_UPLOADER_IP_ADDR;
+import static azkaban.Constants.EventReporterConstants.PROJECT_FILE_UPLOAD_TIME;
+import static azkaban.Constants.EventReporterConstants.PROJECT_FILE_UPLOAD_USER;
+import static azkaban.Constants.EventReporterConstants.PROJECT_NAME;
+import static azkaban.Constants.EventReporterConstants.QUEUE_DURATION;
+import static azkaban.Constants.EventReporterConstants.SLA_OPTIONS;
+import static azkaban.Constants.EventReporterConstants.START_TIME;
+import static azkaban.Constants.EventReporterConstants.SUBMIT_TIME;
+import static azkaban.Constants.EventReporterConstants.SUBMIT_USER;
+import static azkaban.Constants.EventReporterConstants.VERSION;
+import static azkaban.Constants.EventReporterConstants.VERSION_SET;
 import static azkaban.execapp.ConditionalWorkflowUtils.FAILED;
 import static azkaban.execapp.ConditionalWorkflowUtils.PENDING;
 import static azkaban.execapp.ConditionalWorkflowUtils.checkConditionOnJobStatus;
